requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.6.1
neo4j==5.25.0
javalang==0.13.0
GitPython==3.1.43
python-dotenv==1.0.1
rapidfuzz==3.9.7
rich==13.9.4
